succesfully moved object generation into separate service imported into main canvas component
