+++ threejs-ng-snake/src/app/services/obj-builder.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ObjBuilderService } from './obj-builder.service';
+
+describe('ObjBuilderService', () => {
+  let service: ObjBuilderService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ObjBuilderService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
